refactor(redact): thin command controllers, move logic into lib/ (#204)
